+++ backend.php
@@ -25,12 +25,6 @@ if ($_GET['action'] == 'stopvideo') {
 }
 
 
-if ($_GET['action'] == 'stopslaves') {
-	$outputtext =  "not coded yet";
-	//system ("sudo /var/www/sync/stopvideo > /dev/null 2>&1");
-	// this will broadcast a stopall command over tcp or udp
-}
-
 ///////////////////
 //# Video Section
 /////////////////////
@@ -1594,7 +1588,6 @@ if ($_GET['action'] == 'rentalreset') {
 	$outputtext =  "Rental reset done";
 }
 
-
 if ($_GET['action'] == 'installpygame') {
 	$outputtext =  "Install pygame";
 	system("sudo apt-get install -f /var/www/sync/debs/python3-pygame_1.9.3+dfsg-2_armhf.deb");
@@ -1602,7 +1595,7 @@ if ($_GET['action'] == 'installpygame') {
 	$outputtext = "installed pygame";
 }
 
-} ($_GET['action'] == 'updatekernel') {
+if ($_GET['action'] == 'updatekernel') {
 	$outputtext =  "Kernel to RT Kernel, untested yet!";
 	system("sudo cp /var/www/sync/raspberrypi4-linux-4.19.65-rt24-v7l+-usb-lowlatency.tar.gz /raspberrypi4-linux-4.19.65-rt24-v7l+-usb-lowlatency.tar.gz");
 	system("sudo tar -xzf /raspberrypi4-linux-4.19.65-rt24-v7l+-usb-lowlatency.tar.gz");
@@ -2536,6 +2529,7 @@ if ($_GET['action'] == 'stopwebcamusb') {
 
 if ($_GET['action'] == 'onscreenwebcamusb') {
 	$outputtext =  "show webcam on screen/display";
+	system("sudo /var/www/sync/stopall");
 	system("sudo /usr/bin/python /var/www/sync/camerastream.py");
 }