+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # PocketVJ Exhibition / Control Panel 3.x.x
 
+## CP 4.0.3 (not released yet) 21. December 2020 <br />
+
+- added "webcam on screen" display <br />
+- added "install pygame" for countdown and webcam  <br />
+
 ## CP 4.0.2a 18. December 2020 <br />
 
 - fixed missleading rj45 ip when reverting from dhcp  <br />